Add CarouselDemo component for carousel functionality
